Improved fit of the Z peak
The fit is first performed on the MC, using the theoretical natural width. That way, we determine the experimental resolution (Gaussian width). This can then be used in the data fit (not done by default).
The fit is better without constraint, but the two contributions to the width are correlated at 80%. By fixing the experimental resolution, one gets a better estimate of the natural width.